Moved views back into app.py
